# Conflicts:
#	closet/src/main/resources/templates/top/top.html

+++ closet/src/main/java/com/example/model/Coordinate.java
@@ -54,7 +54,7 @@ public class Coordinate {
 	
 	@JsonIgnore
 	@Column(name = "have_calendar")
-	private byte haveCalender;
+	private int haveCalender;
 	
 	/*カレンダー連携*/
 	@JsonIgnore
